kevin/teams_bot/requirements.txt
botbuilder-core>=4.16.1
botbuilder-integration-aiohttp>=4.16.1
aiohttp>=3.9.0
python-dotenv>=1.0.0
